Separate the Hosted Service Client from VM Client
This allows for a less confusing creation of empty cloud services (if
for example one is trying to provision multiple virtual machines into
the same cloud service and wants to create it first).

Also allow specification of the Reverse DNS FQDN for the cloud service
upon creation (will be ommitted from the request to Azure if it's empty
as per the documentation).
